// src/api/generated/models/paginatedFeedingDaysResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { FeedingDayResponseDto } from './feedingDayResponseDto';

export interface PaginatedFeedingDaysResponseDto {
  items: FeedingDayResponseDto[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}