// src/api/generated/models/paginatedFeedingMonthSummariesResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { FeedingMonthSummaryResponseDto } from './feedingMonthSummaryResponseDto';

export interface PaginatedFeedingMonthSummariesResponseDto {
  items: FeedingMonthSummaryResponseDto[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}